refact: Split dashboard store with widget store
